Decompilation checkpoint: 1773079263

+++ src/decomp/sas_c/textdisp2_reset_selection_and_refresh.c
@@ -9,6 +9,6 @@ extern void TEXTDISP2_JMPTBL_ESQIFF_PlayNextExternalAssetFrame(LONG value);
 void TEXTDISP_ResetSelectionAndRefresh(void)
 {
     SCRIPT_UpdateSerialShadowFromCtrlByte(3);
-    TEXTDISP_CurrentMatchIndex = -1;
-    TEXTDISP2_JMPTBL_ESQIFF_PlayNextExternalAssetFrame(0);
+    TEXTDISP_CurrentMatchIndex = (WORD)-1;
+    TEXTDISP2_JMPTBL_ESQIFF_PlayNextExternalAssetFrame((LONG)0);
 }
